requirements.txt
# Core Python Libraries
numpy
pandas
opencv-python
Pillow

# Machine Learning
torch
torchvision
scikit-learn
tensorflow

# Kafka
kafka-python

# Object Storage / Data Lake
minio 

# Dashboard & Visualization
streamlit
plotly

# Time & Date Utilities
tqdm                    
python-dateutil

# Optional (for file watching, etc.)
watchdog
